Extract shared deck.gl helpers from SpatialViewer/Embedding
Pull out parseColor, the bitmask rendering subsystem (LUTs + makeBitmaskLayer factory replacing 3 near-identical layers), and a shared useFocusCellColors hook for the duplicated focus/expression colouring. No behaviour change.
